Implement everything. lol.
This gets things to the point that we have a webcam preview passing
through OpenGL, and a fragment shader generated from an editable node
graph. There aren't any really useful nodes yet (just webcam input and
canvas output, basically), but the things are in place to be able to add
those now.