Redo shader generation, add color blending node.
The new generation setup makes it possible to deal with the shader
build-up immutably (which makes it easier to discard generated shader
elements from nodes that don't actually connect to the final Canvas
node). Nodes take in/output `PartialShader` objects that collect all the
current glsl statements, as well as a working expression (so if a node
defines a variable, the working expression it sends along would be a
GlVarRef to that).

Statements are only added to the final shader if they actually have an
effect on the output (which means we actually write the shader in
reverse... and I probably need to think harder about how that process
works because right now it's convoluted and feels like it could be
simpler).

I may try to add in some things to make the glsl stuff easier to write
for nodes, because right now it's a real pain and super verbose... we'll
see.